requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
astroid=2.3.3=py37_0
ca-certificates=2020.1.1=0
certifi=2019.11.28=py37_0
colorama=0.4.3=py_0
dnspython=1.16.0=pypi_0
isort=4.3.21=py37_0
lazy-object-proxy=1.4.3=py37he774522_0
marshmallow=2.20.5=pypi_0
mccabe=0.6.1=py37_1
motor=2.1.0=pypi_0
openssl=1.1.1d=he774522_4
pip=20.0.2=py37_1
pylint=2.4.4=py37_0
pymongo=3.10.1=pypi_0
python=3.7.6=h60c2a47_2
python-dateutil=2.8.1=pypi_0
setuptools=45.2.0=py37_0
six=1.14.0=py37_0
sqlite=3.31.1=he774522_0
tornado=6.0.3=py37he774522_3
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_1
wheel=0.34.2=py37_0
wincertstore=0.2=py37_0
wrapt=1.11.2=py37he774522_0
